partitioner: Extract token to separate .hh and .cc files
