+++ configure.py
@@ -680,6 +680,7 @@ scylla_core = (['database.cc',
                 'gms/application_state.cc',
                 'gms/inet_address.cc',
                 'dht/i_partitioner.cc',
+                'dht/token.cc',
                 'dht/murmur3_partitioner.cc',
                 'dht/boot_strapper.cc',
                 'dht/range_streamer.cc',
