+++ dht/i_partitioner.cc
@@ -35,41 +35,6 @@
 
 namespace dht {
 
-static const token min_token{ token::kind::before_all_keys, {} };
-static const token max_token{ token::kind::after_all_keys, {} };
-
-const token&
-minimum_token() {
-    return min_token;
-}
-
-const token&
-maximum_token() {
-    return max_token;
-}
-
-int tri_compare(token_view t1, token_view t2) {
-    if (t1._kind < t2._kind) {
-            return -1;
-    } else if (t1._kind > t2._kind) {
-            return 1;
-    } else if (t1._kind == token_kind::key) {
-        return global_partitioner().tri_compare(t1, t2);
-    }
-    return 0;
-}
-
-std::ostream& operator<<(std::ostream& out, const token& t) {
-    if (t._kind == token::kind::after_all_keys) {
-        out << "maximum token";
-    } else if (t._kind == token::kind::before_all_keys) {
-        out << "minimum token";
-    } else {
-        out << global_partitioner().to_sstring(t);
-    }
-    return out;
-}
-
 std::ostream& operator<<(std::ostream& out, const decorated_key& dk) {
     return out << "{key: " << dk._key << ", token:" << dk._token << "}";
 }
